% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icdgenerator.R
\name{generate_sample}
\alias{generate_sample}
\title{Generate Random ICD Data}
\usage{
generate_sample(version = 10, nrows = 1, dcols = 1, pcols = 1,
  gcols = 0, pct_empty = 0.2)
}
\arguments{
\item{version}{ICD version 9 or 10.  Defaults to 10.}

\item{nrows}{number of rows to generate}

\item{dcols}{number of diagnostic codes to generate for each row}

\item{pcols}{number of procedure codes to generate for each row}

\item{gcols}{number of "other" columns to generate}

\item{pct_empty}{the percentage of diagnostic and procedure codes to be
"missing" in the resulting data frame.}
}
\description{
Create a data frame using randomly selected ICD diagnosis and procedure codes
}
\examples{

eg <- generate_sample(
  version = 9,
  nrows = 340, 
  dcols = 29,
  pcols = 15,
  gcols = 114
)

head(eg)
}
