% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-dataset.R
\docType{data}
\name{icd10dx}
\alias{icd10dx}
\title{ICD-10 Diagnostic Codes}
\format{a data frame with 71,704 rows and 2 columns.
\describe{
\item{code}{A character string of alpha numerics representing the ICD-10
diagnostic code.}
\item{long}{Long form label for the ICD-10 diagnostic code}
}}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2018-ICD-10-Code-Descriptions.zip}
}
\usage{
icd10dx
}
\description{
2018 ICD-10 Diagnostic Codes
}
\seealso{
vignette("icd-datasets", package = "icdgenerator") for details
on the data sources and construction of the data set.
}
\keyword{datasets}
