% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-dataset.R
\docType{data}
\name{icd10pc}
\alias{icd10pc}
\title{ICD-10 Procedure Codes}
\format{a data frame with 78,705 rows and 2 columns.
\describe{
\item{code}{A character string of alpha numerics representing the ICD-10
procedure code.}
\item{long}{Long form label for the ICD-10 procedure code}
}}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD10/Downloads/2018-ICD-10-PCS-Codes-File.zip}
}
\usage{
icd10pc
}
\description{
2018 ICD-10 Procedure Codes
}
\seealso{
vignette("icd-datasets", package = "icdgenerator") for details
on the data sources and construction of the data set.
}
\keyword{datasets}
