% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd-dataset.R
\docType{data}
\name{icd9pc}
\alias{icd9pc}
\title{ICD-9-CM Procedure Codes and Labels}
\format{a data frame with 3,882 rows and 3 columns.
\describe{
\item{code}{A character string of alpha numerics representing the ICD-9
procedure code.}
\item{long}{Long form label for the ICD-9 procedure code}
\item{short}{short (abbreviated) form label for the ICD-9 procedure code}
}}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD9ProviderDiagnosticCodes/Downloads/ICD-9-CM-v32-master-descriptions.zip}
}
\usage{
icd9pc
}
\description{
ICD-9-CM Procedure Codes and Labels
}
\seealso{
vignette("icd-datasets", package = "icdgenerator") for details
on the data sources and construction of the data set.
}
\keyword{datasets}
